Corrected linear correction path
The path for linear-correction-0.cfg in the template is relative to the local directory, and refers to an 'extruder' directory, not 'extruders'.
Since the template should be copied into the local printer config, I updated the path to include the repo name, like the other paths specified in the template.